requirements.txt
# python 3.10
importlib-metadata<5.0
flake8
python-dotenv>=0.5.1
pytest
econml
numpy==1.24.2
pandas==2.2.3
pillow==10.4.0
seaborn
scipy==1.14.1
matplotlib
torch==1.12.1
pytorch-lightning==1.7.1
torchmetrics==0.11.4
hydra-core
hydra_colorlog
mlflow
scikit-learn
tables
ray[tune]
torchcde